pyproject.toml
[tool.poetry]
name = "developerglobalchat"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.2"
flask = "^2.0.0"
asyncpg = "^0.22.0"
aiosqlite3 = "^0.3.0"
jishaku = {git = "https://github.com/Gorialis/jishaku.git"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
